[chore] : folder 패키지정리
